A couple of scripts that helped me stress test stores
One is designed to be run in multiple terminals, many that "increment",
and one that "observes", they all run non-stop. Pressing Enter in any
of them shows you their current state. This is the test, basically, because
that forces it to read the files, and make sure you don't get weird
errors on the offsets.

Also, that the "observer" roughly gets the right numbers (the sum of the others)

The other basically increments different labels every time, making the
store grow non-stop. For file and MMap store, there's some logic around
"the file is full now, we need to resize it". This script stresses that,
which otherwise *never* ever gets called, because 1Mb is enough for everybody.

This is how I found all the segfaults...

NOTE: They may have stuff commented out or crap logging added, they're in
whatever state they were when I last used them to actually diagnose something,
not necessarily in super usable state.