Remove other stores
This repository is a fork from the experimental data stores for
prometheus-client available from

https://github.com/gocardless/prometheus-client-ruby-data-stores-experiments

This is only to support redis sotre and remove all other type of stores
and their dependencies to make it possible to use Redis as the backend.

The main reason to go down this path is that the original project
depends on the `mmap2` gem which was only available for Ruby2. So
removing everything other than the redis store seem to a good idea to
move forward for other versions of Ruby.